+++ CHANGELOG.md
@@ -4,7 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/) and [Keep A Ch
 
 ## [Unreleased]
 
-TODO: add at least one Added, Changed, Deprecated, Removed, Fixed or Security section
+### Added
+- 'ScrollingMixin' mixin to scroll elements into the visible viewport of scrollable areas
 
 ## [2.18.2] 2017-08-31
 ### Added