+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/) and [Keep A Changelog's Format](http://keepachangelog.com/).
 
-## [Unreleased]
+## [2.19.0] 2017-09-06
 
 ### Added
 - 'ScrollingMixin' mixin to scroll elements into the visible viewport of scrollable areas